requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
requests
python-dotenv
email-validator
python-whois
cryptography
ipwhois
dnspython
APScheduler
WeasyPrint
tldextract
